backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==42.0.8
dj-database-url==1.3.0
Django==5.0.14
django-cors-headers==4.3.1
django-encrypted-model-fields==0.6.5
django-filter==24.3
django-ipware==7.0.1
django-structlog==10.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==3.0.0
redis==5.3.1
requests==2.31.0
sentry-sdk==1.40.6
six==1.17.0
sqlparse==0.5.5
structlog==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.6.0
gunicorn==23.0.0
